Build fixes.

+++ onnxruntime/contrib_ops/cpu/attnlstm/deep_cpu_attn_lstm.h
@@ -96,7 +96,7 @@ class DeepCpuAttnLstmOp final : public OpKernel {
 // across them. mutable due to this.
 // The alternative would be to create a threadpool in each call to Compute but that would incur thread creation
 // cost on every call.
-  mutable onnxruntime::concurrency::ThreadPool ttp_{"DEEPCPU_ATTN_LSTM", std::thread::hardware_concurrency()};
+  mutable onnxruntime::concurrency::ThreadPool ttp_{"DEEPCPU_ATTN_LSTM", (int)std::thread::hardware_concurrency()};
 };
 
 }  // namespace contrib
